+++ src/components/Checkbox/index.tsx
@@ -95,13 +95,11 @@ export class CheckboxInt extends React.PureComponent<CheckboxProps, CheckboxStat
     };
   }
 
-  componentWillReceiveProps(nextProps: CheckboxProps) {
+  componentWillReceiveProps({ value = false, error }: CheckboxProps) {
     if (this.state.controlled) {
-      this.setState({
-        value: nextProps.value || false,
-        error: nextProps.error,
-      });
+      this.setState({ value });
     }
+    this.setState({ error });
   }
 
   componentDidMount() {
@@ -194,7 +192,17 @@ export class CheckboxInt extends React.PureComponent<CheckboxProps, CheckboxStat
   };
 
   render() {
-    const { children, disabled, theme, value: _0, defaultValue: _1, onChange: _2, info, ...props } = this.props;
+    const {
+      children,
+      disabled,
+      theme,
+      value: _0,
+      defaultValue: _1,
+      onChange: _2,
+      onInput: _3,
+      info,
+      ...props
+    } = this.props;
     const { value, error } = this.state;
     const containerProps = {
       ...props,