+++ src/components/DropdownField/DropdownFieldInt.tsx
@@ -17,6 +17,7 @@ import {
   InteractiveListWrapperProps,
   InteractiveListDirection,
   InteractiveListBorderType,
+  InteractiveListProps,
 } from '../InteractiveList';
 import {
   StyledInputRow,
@@ -217,7 +218,7 @@ export class DropdownFieldInt extends React.Component<DropdownFieldProps & FormC
 
     this.state = {
       value: getIndices(data, value, props.multiple),
-      open: false,
+      open: props.open === true,
       controlled: props.value !== undefined,
       error: props.error,
     };
@@ -239,17 +240,16 @@ export class DropdownFieldInt extends React.Component<DropdownFieldProps & FormC
     }
   }
 
-  componentWillReceiveProps(nextProps: DropdownFieldProps) {
+  componentWillReceiveProps({ data = [], value = [], error, multiple }: DropdownFieldProps) {
     const { controlled } = this.state;
 
     if (controlled) {
-      const data = nextProps.data || [];
-      const value = nextProps.value !== undefined ? nextProps.value : [];
       this.setState({
-        value: getIndices(data, value, nextProps.multiple),
-        error: nextProps.error,
+        value: getIndices(data, value, multiple),
       });
     }
+
+    this.setState({ error });
   }
 
   private show = () =>
@@ -338,7 +338,7 @@ export class DropdownFieldInt extends React.Component<DropdownFieldProps & FormC
   };
 
   private renderList = (screenSize?: ScreenSize) => {
-    const { data = [], theme, disabled, multiple } = this.props;
+    const { data = [], theme, disabled, multiple, direction } = this.props;
     const { open, value } = this.state;
     const mobile = screenSize === 'small';
     const wrapper = mobile ? getMobileWrapper(<StyledLabel>{this.props.label}</StyledLabel>) : StandardWrapper;
@@ -356,7 +356,8 @@ export class DropdownFieldInt extends React.Component<DropdownFieldProps & FormC
         indices={value}
         customWrapper={wrapper}
         onClickOutside={() => {}}
-        autoPosition
+        direction={direction}
+        autoPosition={undefined === direction}
         autoFocus
       />
     );
@@ -368,6 +369,7 @@ export class DropdownFieldInt extends React.Component<DropdownFieldProps & FormC
       children: _0,
       value: _1,
       defaultValue: _2,
+      onInput: _3,
       data = [],
       theme,
       className: classNameProp,