+++ src/components/Sidebar/SidebarActivatorContainer.part.tsx
@@ -1,0 +1,3 @@
+import styled from '../../utils/styled';
+
+export const SidebarActivatorContainer = styled.div``;